Transforma tela de cadastrar novo em um popup para padronização do projeto
